requirements.txt
locustio==0.999
pytest==7.4.0
Requests==2.31.0
pytest-html==4.0.0rc5
pytest-xdist==3.3.1
